partially fixes resource dashboard
