+++ main/app/Resources/translations/actions.en.json
@@ -66,6 +66,7 @@
   "show-calendar": "Show calendar",
   "show-colors": "Show color chart",
   "show-comments": "Show comments",
+  "show-dashboard": "Show dashboard",
   "show-followers": "Show followers",
   "show-help": "Show help",
   "show-info": "Show information",
