+++ main/app/Resources/translations/actions.fr.json
@@ -66,6 +66,7 @@
   "show-calendar": "Voir le calendrier",
   "show-colors": "Voir le nuancier",
   "show-comments": "Voir les commentaires",
+  "show-dashboard": "Voir le tableau de bord",
   "show-followers": "Voir les abonnés",
   "show-help": "Voir l'aide",
   "show-info": "Voir les informations",
