+++ main/core/Resources/modules/resource/actions/dashboard.js
@@ -4,10 +4,10 @@ import {LINK_BUTTON} from '#/main/app/buttons'
 /**
  * Opens the dashboard page of a resource.
  */
-export default () => ({
+export default (resourceNodes, nodesRefresher, path) => ({
   name: 'dashboard',
   type: LINK_BUTTON,
   icon: 'fa fa-fw fa-tachometer',
-  label: trans('dashboard', {}, 'tools'),
-  target: '/dashboard'
+  label: trans('show-dashboard', {}, 'actions'),
+  target: `${path}/${resourceNodes[0].slug}/dashboard`
 })
