+++ main/core/Resources/modules/resource/dashboard/containers/connections.jsx
@@ -3,7 +3,7 @@ import {connect} from 'react-redux'
 
 import {selectors as resourceSelectors} from '#/main/core/resource/store'
 
-import {Connections as ConnectionsComponent} from '#/plugin/path/resources/path/dashboard/components/connections'
+import {Connections as ConnectionsComponent} from '#/main/core/resource/dashboard/components/connections'
 
 const Connections = connect(
   state => ({
