+++ main/core/Resources/modules/resource/dashboard/containers/logs-user.jsx
@@ -3,7 +3,7 @@ import {connect} from 'react-redux'
 
 import {selectors as resourceSelectors} from '#/main/core/resource/store'
 
-import {UserLogs as UserLogsComponent} from '#/plugin/path/resources/path/dashboard/components/logs-user'
+import {UserLogs as UserLogsComponent} from '#/main/core/resource/dashboard/components/logs-user'
 
 const UserLogs = connect(
   state => ({
