+++ main/core/Resources/modules/resource/dashboard/containers/main.jsx
@@ -1,0 +1,26 @@
+import {connect} from 'react-redux'
+
+import {withReducer} from '#/main/app/store/components/withReducer'
+import {selectors as resourceSelectors} from '#/main/core/resource/store'
+import {actions as logActions} from  '#/main/core/layout/logs/actions'
+
+import {DashboardMain as DashboardMainComponent} from '#/main/core/resource/dashboard/components/main'
+import {actions, reducer, selectors} from '#/main/core/resource/dashboard/store'
+
+const DashboardMain = withReducer(selectors.STORE_NAME, reducer)(
+  connect(
+    (state) => ({
+      path: resourceSelectors.path(state),
+      resourceId: resourceSelectors.resourceNode(state).autoId
+    }),
+    dispatch => ({
+      openLog(id, resourceId) {
+        dispatch(logActions.openLog('apiv2_resource_logs_get', {id, resourceId}))
+      }
+    })
+  )(DashboardMainComponent)
+)
+
+export {
+  DashboardMain
+}
