+++ main/core/Resources/modules/resource/dashboard/containers/progression.jsx
@@ -1,7 +1,7 @@
 import {connect} from 'react-redux'
 
 import {selectors as pathSelectors} from '#/plugin/path/resources/path/store'
-import {Progression as ProgressionComponent} from '#/plugin/path/resources/path/dashboard/components/progression'
+import {Progression as ProgressionComponent} from '#/main/core/resource/dashboard/components/progression'
 
 const Progression = connect(
   (state) => ({
