+++ main/core/Resources/modules/resource/dashboard/store/actions.js
