+++ main/core/Resources/modules/resource/dashboard/store/index.js
@@ -1,0 +1,13 @@
+/**
+ * Path dashboard store.
+ */
+
+import {actions}   from '#/main/core/resource/dashboard/store/actions'
+import {reducer}   from '#/main/core/resource/dashboard/store/reducer'
+import {selectors}   from '#/main/core/resource/dashboard/store/selectors'
+
+export {
+  actions,
+  reducer,
+  selectors
+}
