--- main/core/Resources/modules/resource/logs/connection/store/index.js
@@ -1,5 +1,0 @@
-import {reducer} from '#/main/core/resource/logs/connection/store/reducer'
-
-export {
-  reducer
-}
