+++ plugin/path/Resources/modules/resources/path/components/resource.jsx
@@ -8,7 +8,6 @@ import {ResourcePage} from '#/main/core/resource/containers/page'
 import {OverviewMain} from '#/plugin/path/resources/path/overview/containers/main'
 import {EditorMain} from '#/plugin/path/resources/path/editor/containers/main'
 import {PlayerMain} from '#/plugin/path/resources/path/player/containers/main'
-import {DashboardMain} from '#/plugin/path/resources/path/dashboard/containers/main'
 
 const PathResource = props =>
   <ResourcePage
@@ -25,12 +24,6 @@ const PathResource = props =>
         icon: 'fa fa-fw fa-play',
         label: trans('start', {}, 'actions'),
         target: `${props.path}/play`
-      }, {
-        type: LINK_BUTTON,
-        icon: 'fa fa-fw fa-tachometer',
-        label: trans('dashboard'),
-        displayed: props.editable,
-        target: `${props.path}/dashboard`
       }
     ]}
     routes={[
@@ -46,10 +39,6 @@ const PathResource = props =>
         exact: true,
         component: OverviewMain,
         disabled: !props.overview
-      }, {
-        path: '/dashboard',
-        component: DashboardMain,
-        disabled: !props.editable
       }
     ]}
     redirect={[
