--- plugin/path/Resources/modules/resources/path/dashboard/store/index.js
@@ -1,13 +1,0 @@
-/**
- * Path dashboard store.
- */
-
-import {actions}   from '#/plugin/path/resources/path/dashboard/store/actions'
-import {reducer}   from '#/plugin/path/resources/path/dashboard/store/reducer'
-import {selectors}   from '#/plugin/path/resources/path/dashboard/store/selectors'
-
-export {
-  actions,
-  reducer,
-  selectors
-}
